Support for GitHub enterprise
